backend/database.py
"""
PostgreSQL Database configuration using SQLAlchemy.
Reads DATABASE_URL from .env file.
Fallback: SQLite for local dev if no PostgreSQL URL is set.
"""
import os
from sqlalchemy import create_engine
from sqlalchemy.ext.declarative import declarative_base
from sqlalchemy.orm import sessionmaker
from dotenv import load_dotenv

load_dotenv()

DATABASE_URL = os.getenv(
    "DATABASE_URL",
    "sqlite:///./mediai_appointments.db"  # Fallback for local dev
)

# PostgreSQL requires no additional args; SQLite requires check_same_thread
connect_args = {}
if DATABASE_URL.startswith("sqlite"):
    connect_args = {"check_same_thread": False}

engine = create_engine(DATABASE_URL, connect_args=connect_args)

SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)

Base = declarative_base()


def get_db():
    """Dependency to get a DB session."""
    db = SessionLocal()
    try:
        yield db
    finally:
        db.close()


def init_db():
    """Create all tables if they don't exist."""
    from models import Appointment, PatientSession  # noqa: F401
    Base.metadata.create_all(bind=engine)
    print(f"[DB] Database initialized: {DATABASE_URL.split('@')[-1] if '@' in DATABASE_URL else DATABASE_URL}")
