backend/requirements.txt
fastapi[standard]
uvicorn
chromadb
sentence-transformers
pydantic
python-dotenv
requests
openai
